Updating data in an EFRepository using tracking instead of a Update

+++ UniversityProject/UCore/MillitaryClass.cs
@@ -2,7 +2,7 @@
 
 public class MillitaryClass
 {
-    public int MillitaryId { get; set; }
+    public long MillitaryId { get; set; }
     public string LevelId { get; set; }
     public ICollection<Student> Student { get; set; } = new List<Student>();
 }
