+++ UniversityProject/UCore/Student.cs
@@ -7,7 +7,7 @@ using Logger;
 public class Student
 {
     public long PassportId { get; set; }
-    public int MillitaryId { get; set; }
+    public long MillitaryId { get; set; }
     public MillitaryClass Millitary { get; set; }
     public Passport Passport { get; set; }
     public bool CriminalRecord { get; set; }
